+++ src/cpp/tests/test_abaqus_tools.cpp
@@ -151,17 +151,17 @@ BOOST_AUTO_TEST_CASE( testContractAbaqusNTENSMatrix ){
      //Test full size matrix
      std::vector< std::vector< double > > matrix_contraction_full( 6, std::vector< double >( 6, -666. ) );
      std::vector< std::vector< double > > abaqus_full   = { { 1111, 1122, 1133, 1112, 1113, 1123 },
-                                                            { 1122, 2222, 2233, 2212, 2213, 2223 },
-                                                            { 1133, 2233, 3333, 3312, 3313, 3323 },
-                                                            { 1112, 2212, 3312, 1212, 1213, 1223 },
-                                                            { 1113, 2213, 3313, 1213, 1313, 1323 },
-                                                            { 1123, 2223, 3323, 1223, 1323, 2323 } };
+                                                            { 2211, 2222, 2233, 2212, 2213, 2223 },
+                                                            { 3311, 3322, 3333, 3312, 3313, 3323 },
+                                                            { 1211, 1222, 1233, 1212, 1213, 1223 },
+                                                            { 1311, 1322, 1333, 1312, 1313, 1323 },
+                                                            { 2311, 2322, 2333, 2312, 2313, 2323 } };
      std::vector< std::vector< double > > expanded_full = { { 1111, 1122, 1133, 1112, 1113, 1123 },
-                                                            { 1122, 2222, 2233, 2212, 2213, 2223 },
-                                                            { 1133, 2233, 3333, 3312, 3313, 3323 },
-                                                            { 1112, 2212, 3312, 1212, 1213, 1223 },
-                                                            { 1113, 2213, 3313, 1213, 1313, 1323 },
-                                                            { 1123, 2223, 3323, 1223, 1323, 2323 } };
+                                                            { 2211, 2222, 2233, 2212, 2213, 2223 },
+                                                            { 3311, 3322, 3333, 3312, 3313, 3323 },
+                                                            { 1211, 1222, 1233, 1212, 1213, 1223 },
+                                                            { 1311, 1322, 1333, 1312, 1313, 1323 },
+                                                            { 2311, 2322, 2333, 2312, 2313, 2323 } };
      NDI = 3;
      NSHR = 3;
 
@@ -172,12 +172,12 @@ BOOST_AUTO_TEST_CASE( testContractAbaqusNTENSMatrix ){
      //Test plane stress matrix
      std::vector< std::vector< double > > matrix_contraction_plane_stress( 3, std::vector< double >( 3, -666. ) );
      std::vector< std::vector< double > > abaqus_plane_stress   = { { 1111, 1122, 1112 },
-                                                                    { 1122, 2222, 2212 },
-                                                                    { 1112, 2212, 1212 } };
+                                                                    { 2211, 2222, 2212 },
+                                                                    { 1211, 1222, 1212 } };
      std::vector< std::vector< double > > expanded_plane_stress = { { 1111, 1122, 0.00, 1112, 0.00, 0.00 },
-                                                                    { 1122, 2222, 0.00, 2212, 0.00, 0.00 },
+                                                                    { 2211, 2222, 0.00, 2212, 0.00, 0.00 },
                                                                     { 0.00, 0.00, 0.00, 0.00, 0.00, 0.00 },
-                                                                    { 1112, 2212, 0.00, 1212, 0.00, 0.00 },
+                                                                    { 1211, 1222, 0.00, 1212, 0.00, 0.00 },
                                                                     { 0.00, 0.00, 0.00, 0.00, 0.00, 0.00 },
                                                                     { 0.00, 0.00, 0.00, 0.00, 0.00, 0.00 } };
      NDI = 2;
@@ -186,5 +186,100 @@ BOOST_AUTO_TEST_CASE( testContractAbaqusNTENSMatrix ){
      matrix_contraction_plane_stress = abaqusTools::contractAbaqusNTENSMatrix( expanded_plane_stress, NDI, NSHR );
 
      BOOST_CHECK( vectorTools::fuzzyEquals( matrix_contraction_plane_stress, abaqus_plane_stress ) );
+
+}
+
+BOOST_AUTO_TEST_CASE( testExpandFullNTENSTensor ){
+    /*!
+     * Test expansion of full stress and strain type tensors from full Abaqus vectors
+     */
+
+     //Initialize common test variables
+     std::vector< int > result( 9 );
+     std::vector< int > abaqus_standard = { 11, 22, 33, 12, 13, 23 };
+     std::vector< int > abaqus_explicit = { 11, 22, 33, 12, 23, 13 };
+     std::vector< int > expected = { 11, 12, 13,
+                                     12, 22, 23,
+                                     13, 23, 33 };
+
+     //Check full tensor expansion for Abaqus/Standard
+     result = abaqusTools::expandFullNTENSTensor( abaqus_standard, true );
+     BOOST_CHECK( result == expected );
+
+     //Check full tensor expansion for Abaqus/Explicit
+     result = abaqusTools::expandFullNTENSTensor( abaqus_explicit, false );
+     BOOST_CHECK( result == expected );
+
+     //Test overloaded variation with contracted vector input
+     //Check full tensor expansion for Abaqus/Standard
+     result = abaqusTools::expandFullNTENSTensor( abaqus_standard, 3, 3, true );
+     BOOST_CHECK( result == expected );
+
+     //Check full tensor expansion for Abaqus/Explicit
+     result = abaqusTools::expandFullNTENSTensor( abaqus_explicit, 3, 3, false );
+     BOOST_CHECK( result == expected );
+
+}
+
+BOOST_AUTO_TEST_CASE( testContractFullNTENSTensor ){
+
+     //Initialize common test variables
+     std::vector< int > result( 6 );
+     std::vector< int > full_tensor = { 11, 12, 13,
+                                        12, 22, 23,
+                                        13, 23, 33 };
+     std::vector< int > abaqus_standard = { 11, 22, 33, 12, 13, 23 };
+     std::vector< int > abaqus_explicit = { 11, 22, 33, 12, 23, 13 };
+
+     //Check full tensor contraction for Abaqus/Standard
+     result = abaqusTools::contractFullNTENSTensor( full_tensor, true );
+     BOOST_CHECK( result == abaqus_standard );
+
+     //Check full tensor contraction for Abaqus/Explicit
+     result = abaqusTools::contractFullNTENSTensor( full_tensor, false );
+     BOOST_CHECK( result == abaqus_explicit );
+
+     //Test overloaded variation with contracted vector input
+     //Check full tensor contraction for Abaqus/Standard
+     result = abaqusTools::contractFullNTENSTensor( full_tensor, 3, 3, true );
+     BOOST_CHECK( result == abaqus_standard );
+
+     //Check full tensor contraction for Abaqus/Explicit
+     result = abaqusTools::contractFullNTENSTensor( full_tensor, 3, 3, false );
+     BOOST_CHECK( result == abaqus_explicit );
+
+}
+
+BOOST_AUTO_TEST_CASE( testContractFullNTENSMatrix ){
+
+     //Initialize common test variables
+     std::vector< std::vector< int > > result( 6, std::vector< int >( 6 ) );
+                                                         // 0     1     2     3     4     5     6     7     8
+     std::vector< std::vector< int > > full_matrix = { { 1111, 1112, 1113, 1121, 1122, 1123, 1131, 1132, 1133 },   // 0
+                                                       { 1211, 1212, 1213, 1221, 1222, 1223, 1231, 1232, 1233 },   // 1
+                                                       { 1311, 1312, 1313, 1321, 1322, 1323, 1331, 1332, 1333 },   // 2
+                                                       { 2111, 2112, 2113, 2121, 2122, 2123, 2131, 2132, 2133 },   // 3
+                                                       { 2211, 2212, 2213, 2221, 2222, 2223, 2231, 2232, 2233 },   // 4
+                                                       { 2311, 2312, 2313, 2321, 2322, 2323, 2331, 2332, 2333 },   // 5
+                                                       { 3111, 3112, 3113, 3121, 3122, 3123, 3131, 3132, 3133 },   // 6
+                                                       { 3211, 3212, 3213, 3221, 3222, 3223, 3231, 3232, 3233 },   // 7
+                                                       { 3311, 3312, 3313, 3321, 3322, 3323, 3331, 3332, 3333 } }; // 8
+                                                                        //  0     4     8     1     2     5
+     std::vector< std::vector< int > > full_abaqus_standard_matrix = { { 1111, 1122, 1133, 1112, 1113, 1123 },   // 0
+                                                                       { 2211, 2222, 2233, 2212, 2213, 2223 },   // 4
+                                                                       { 3311, 3322, 3333, 3312, 3313, 3323 },   // 8
+                                                                       { 1211, 1222, 1233, 1212, 1213, 1223 },   // 1
+                                                                       { 1311, 1322, 1333, 1312, 1313, 1323 },   // 2
+                                                                       { 2311, 2322, 2333, 2312, 2313, 2323 } }; // 5
+
+     //Check full matrix contraction for Abaqus/Standard
+     result = abaqusTools::contractFullNTENSMatrix( full_matrix );
+     BOOST_CHECK( result == full_abaqus_standard_matrix );
+
+     //Test overloaded variation with contracted NTENSxNTENS matrix output
+     //Check full matrix contraction for Abaqus/Standard
+     //Sign of life test. The tests for NDI and NSHR less than 3 are performed in contractAbaqusNTENSMatrix.
+     result = abaqusTools::contractFullNTENSMatrix( full_matrix, 3, 3 );
+     BOOST_CHECK( result == full_abaqus_standard_matrix );
 
 }